+++ index.d.ts
@@ -141,7 +141,7 @@ declare module 'react-pixi-fiber' {
 
   /** `Stage` component properties." */
   export interface StageProperties extends Component<PIXI.Container & InteractiveComponent> {
-    options?: PIXI.ApplicationOptions
+    options?: {}
   }
 
   /**